chore(af-sui-types): remove `Owner` TODOs
We can't make the changes in these TODOs because `sui-jsonrpc` is tied
to the serialization format of the current `Owner`. For instance, a
transaction response from JSON-RPC will have
```json
{
  "digest": "FWGKbaHPH7yUtXJrDr42FSdEL9P2UZCR2zqmQypiaEni",
  "effects": {
    "gasObject": {
      "owner": {
        "AddressOwner": "0x76f9ca7f89994d4039b739859a41b39123d6f695a1b33f7431cee3c6b40a45c2"
      },
      "reference": {
        "objectId": "0x01cd3c322c2f8ff03850677984662862cc2822c8e9c2612cad5b25a27e6fa5bd",
        "version": 185355111,
        "digest": "Epg2DHeHD7RMSGJyYEEddmsXkxv47NJXCZt1aEB4NEYY"
      }
    },
    ...
  }
}
```
The `"AddressOwner"` tag above is derived from the current
`Owner`. Thus, we'd need to introduce alternative serializations
for `sui-sdk-types::Owner` before replacing our own with it.