refactor: make public dependencies explicit
This leverages [RFC 3516] to mark dependencies as public if that's the
case.

It makes the job of figuring out 'should I bump the major version of my
crate if I bumped the major version of dependency X' almost trivial.
This has been a fairly tedious and error-prone endeavour in the past, as
it required 'scanning' the public API of a crate every time a dependency
was updated to see if one of its items was exposed.

Its is accompanied by a [lint], enabled in `.cargo/config.toml`. Mind
you, it [isn't perfect], but goes a **long** way in avoiding a lot
common pitfalls.

I found this after doing more research on public dependencies since
managing them was one of the maintenance tasks I took on. Don't
underestimate this issue, as mishandling versioning in these scenarions
leads to compilation errors and lots of time wasted 'patching' crates.

[RFC 3516]: https://rust-lang.github.io/rfcs/3516-public-private-dependencies.html
[lint]: https://doc.rust-lang.org/nightly/rustc/lints/listing/warn-by-default.html#exported-private-dependencies
[isn't perfect]: https://github.com/rust-lang/rust/issues/71043